Working on csvExportActions.test to get fs mocking working together with git mocking.
