+++ __tests__/ProjectDetailsActions.test.js
@@ -1,7 +1,6 @@
 /* eslint-env jest */
 import fs from 'fs-extra';
 import path from 'path-extra';
-import ospath from "ospath";
 import types from '../src/js/actions/ActionTypes';
 import * as actions from '../src/js/actions/ProjectDetailsActions';
 import thunk from 'redux-thunk';
@@ -10,7 +9,6 @@ import configureMockStore from 'redux-mock-store';
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 
-const PROJECTS_PATH = path.join(ospath.home(), 'translationCore', 'projects');
 
 it('setSaveLocation() creates an action to update contributors', () => {
   const store = mockStore({});
@@ -208,181 +206,3 @@ it('updateCheckers() creates an action to update checkers', () => {
   const receivedActions = store.getActions();
   expect(receivedActions).toEqual(expectedActions);
 });
-
-describe('ProjectDetailsActions.updateProjectNameIfNecessary()', () => {
-  const currentProjectName = "fr_ult_eph_book";
-  const currentProjectPath = path.join(PROJECTS_PATH, currentProjectName);
-  const mockStoreData = {
-    projectDetailsReducer: {
-      manifest: {
-        target_language: {
-          id: 'fr',
-          name: 'francais',
-          direction: 'ltr'
-        },
-        project: {
-          id: 'eph',
-          name: 'Ephesians'
-        },
-        resource: {
-          id: 'ult',
-          name: 'unfoldingWord Literal Text'
-        }
-      },
-      projectSaveLocation: currentProjectPath
-    },
-    localImportReducer: {
-      selectedProjectFilename: 'SELECTED_PROJECT_NAME'
-    }
-  };
-
-  beforeEach(() => {
-    // reset mock filesystem data
-    fs.__resetMockFS();
-    // Set up mock filesystem before each test
-    fs.__setMockFS({
-      [currentProjectPath]: ''
-    });
-  });
-
-  afterEach(() => {
-    // reset mock filesystem data
-    fs.__resetMockFS();
-  });
-
-  test('does nothing if project name is valid', async () => {
-    // given
-    const store = mockStore(mockStoreData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).toBeTruthy();
-  });
-
-  test('does nothing if projectSaveLocation is not set', async () => {
-    // given
-    const storeData = JSON.parse(JSON.stringify(mockStoreData));
-    delete storeData.projectDetailsReducer.projectSaveLocation;
-    const store = mockStore(storeData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).toBeTruthy();
-  });
-
-  test('renames project if lang_id changed', async () => {
-    // given
-    const newProjectName = "am_ult_eph_book";
-    const expectedProjectPath = path.join(PROJECTS_PATH, newProjectName);
-    const storeData = JSON.parse(JSON.stringify(mockStoreData));
-    storeData.projectDetailsReducer.manifest.target_language.id = 'am';
-    const store = mockStore(storeData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).not.toBeTruthy();
-    expect(fs.pathExistsSync(expectedProjectPath)).toBeTruthy();
-  });
-
-  test('renames project if project id changed', async () => {
-    // given
-    const newProjectName = "fr_ult_tit_book";
-    const expectedProjectPath = path.join(PROJECTS_PATH, newProjectName);
-    const storeData = JSON.parse(JSON.stringify(mockStoreData));
-    storeData.projectDetailsReducer.manifest.project.id = 'tit';
-    const store = mockStore(storeData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).not.toBeTruthy();
-    expect(fs.pathExistsSync(expectedProjectPath)).toBeTruthy();
-  });
-
-  test('renames project if resource.id (resourceId) changed', async () => {
-    // given
-    const newProjectName = "fr_lib_eph_book";
-    const expectedProjectPath = path.join(PROJECTS_PATH, newProjectName);
-    const storeData = JSON.parse(JSON.stringify(mockStoreData));
-    storeData.projectDetailsReducer.manifest.resource.id = 'lib';
-    const store = mockStore(storeData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).not.toBeTruthy();
-    expect(fs.pathExistsSync(expectedProjectPath)).toBeTruthy();
-  });
-
-  test('renames project if new project name is different than spec', async () => {
-    // given
-    const currentProjectPath = path.join(PROJECTS_PATH, "fr_ULT_eph_book");
-    const newProjectName = "fr_ult_eph_book";
-    const expectedProjectPath = path.join(PROJECTS_PATH, newProjectName);
-    fs.moveSync(expectedProjectPath, currentProjectPath); // move to invalid file
-    const storeData = JSON.parse(JSON.stringify(mockStoreData));
-    storeData.projectDetailsReducer.projectSaveLocation = currentProjectPath;
-    const store = mockStore(storeData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).not.toBeTruthy();
-    expect(fs.pathExistsSync(expectedProjectPath)).toBeTruthy();
-  });
-
-  test('does not rename project if new project name is different than spec and we have duplicate', async () => {
-    // given
-    const currentProjectPath = path.join(PROJECTS_PATH, "fr_ULT_eph_book");
-    const newProjectName = "fr_ult_eph_book";
-    const expectedProjectPath = path.join(PROJECTS_PATH, newProjectName);
-    fs.moveSync(expectedProjectPath, currentProjectPath); // move to invalid file
-    fs.copySync(currentProjectPath, expectedProjectPath); // make duplicate
-    const storeData = JSON.parse(JSON.stringify(mockStoreData));
-    storeData.projectDetailsReducer.projectSaveLocation = currentProjectPath;
-    const store = mockStore(storeData);
-
-    // when
-    await store.dispatch(actions.updateProjectNameIfNecessary());
-
-    // then
-    expect(cleanupPaths(store.getActions())).toMatchSnapshot();
-    expect(fs.pathExistsSync(currentProjectPath)).toBeTruthy();
-    expect(fs.pathExistsSync(expectedProjectPath)).toBeTruthy();
-  });
-});
-
-//
-// helpers
-//
-
-/**
- * remove user specific paths and just get basename
- * @param {Array} actions
- * @return {*}
- */
-function cleanupPaths(actions) {
-  if (actions && actions.length) {
-    for (let action of actions) {
-      if ('pathLocation' in action) {
-        action.pathLocation = path.basename(action.pathLocation);
-      }
-    }
-  }
-  return actions;
-}
